adding requested fixes

+++ src/include/duckdb.h
@@ -47,14 +47,14 @@ typedef enum DUCKDB_TYPE {
 
 typedef struct {
 	int16_t year;
-	int16_t month;
-	int16_t day;
+	int8_t month;
+	int8_t day;
 } duckdb_date;
 
 typedef struct {
-	int16_t hour;
-	int16_t min;
-	int16_t sec;
+	int8_t hour;
+	int8_t min;
+	int8_t sec;
 	int16_t msec;
 } duckdb_time;
